+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Ability to pin on home page without an account
+
+### Removed
+- User profile pages
 
 ## [1.0.0](https://gitlab.com/jrswab/nebulus/commit/b585d3b4137c354e17e930dbb5bb766ebf3dbffc) - 2018-08-27
 ### Added